+++ src/AtmosphericModels.jl
@@ -6,30 +6,25 @@
 
 module AtmosphericModels
 
-using Accessors
-using Crayons
 using Dates
 using LinearAlgebra
 using Printf
-using SpaceIndices
+
+using Accessors
 using SatelliteToolboxBase
 using SatelliteToolboxCelestialBodies
 using SatelliteToolboxLegendre
+using SpaceIndices
 using StaticArraysCore
 
 import Base: show
 
 ############################################################################################
-#                                        Constants                                         #
-############################################################################################
-
-const _D = string(Crayon(; reset = true))
-const _B = string(crayon"bold")
-
-############################################################################################
 #                                         Includes                                         #
 ############################################################################################
-include("utils.jl")
+
+include("./types.jl")
+include("./utils.jl")
 
 include("./exponential/constants.jl")
 include("./exponential/exponential.jl")
@@ -41,23 +36,21 @@ include("./harrispriester/harrispriester_modified.jl")
 include("./jacchia1977/types.jl")
 include("./jacchia1977/constants.jl")
 include("./jacchia1977/jacchia1977.jl")
-include("./jacchia1977/show.jl")
 
 include("./jr1971/types.jl")
 include("./jr1971/constants.jl")
 include("./jr1971/jr1971.jl")
-include("./jr1971/show.jl")
 
 include("./jb2008/types.jl")
 include("./jb2008/constants.jl")
 include("./jb2008/jb2008.jl")
-include("./jb2008/show.jl")
 
 include("./nrlmsise00/types.jl")
 include("./nrlmsise00/auxiliary.jl")
 include("./nrlmsise00/constants.jl")
 include("./nrlmsise00/math.jl")
 include("./nrlmsise00/nrlmsise00.jl")
-include("./nrlmsise00/show.jl")
+
+include("./show.jl")
 
 end # module AtmosphericModels