+++ src/jacchia1977/types.jl
@@ -7,7 +7,7 @@
 export Jacchia1977Output
 
 """
-    struct Jacchia1977Output{T<:Number}
+    struct Jacchia1977Output{T <: Number} <: AbstractAtmosphericModelOutput
 
 Output of the atmospheric model Jacchia 1977.
 
@@ -27,7 +27,7 @@ Output of the atmospheric model Jacchia 1977.
 - `He_number_density::T`: Number density of He [1 / m³].
 - `H_number_density::T`: Number density of H [1 / m³].
 """
-struct Jacchia1977Output{T <: Number}
+struct Jacchia1977Output{T <: Number} <: AbstractAtmosphericModelOutput
     total_density::T
     temperature::T
     exospheric_temperature::T
@@ -37,4 +37,21 @@ struct Jacchia1977Output{T <: Number}
     Ar_number_density::T
     He_number_density::T
     H_number_density::T
+end
+
+_model_name(::Type{<:Jacchia1977Output}) = "Jacchia 1977"
+_model_description(::Type{<:Jacchia1977Output}) = "Jacchia 1977"
+
+function _show_fields(::Type{<:Jacchia1977Output})
+    return (
+        ("Total density", :total_density, _SHOW_FORMAT_DENSITY, "kg / m³"),
+        ("Temperature", :temperature, _SHOW_FORMAT_TEMPERATURE, "K"),
+        ("Exospheric Temp.", :exospheric_temperature, _SHOW_FORMAT_TEMPERATURE, "K"),
+        ("N₂ number density", :N2_number_density, _SHOW_FORMAT_DENSITY, "1 / m³"),
+        ("O₂ number density", :O2_number_density, _SHOW_FORMAT_DENSITY, "1 / m³"),
+        ("O  number density", :O_number_density, _SHOW_FORMAT_DENSITY, "1 / m³"),
+        ("Ar number density", :Ar_number_density, _SHOW_FORMAT_DENSITY, "1 / m³"),
+        ("He number density", :He_number_density, _SHOW_FORMAT_DENSITY, "1 / m³"),
+        ("H  number density", :H_number_density, _SHOW_FORMAT_DENSITY, "1 / m³"),
+    )
 end